# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Youtube Video API",
        "description": "This is a Youtube Video API that allow users to retrieve generated data about youtube video(s) as a JSON format.",
        "contact": {
            "email": "n00201327@iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Display all comments",
                "description": "Displays all the comments with its relationship ie. Channel that made the comment.",
                "operationId": "0406f270b674fae17bf23b4206a3004f",
                "parameters": [
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "orderBy allows data to be re-arrange in the order of title, created_at, views, likes, and dislikes.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "random",
                        "in": "query",
                        "description": "retrieve data and pluck it in random order.",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "likes",
                        "in": "query",
                        "description": "retrieve data by filtering the comments that have the amount of likes less than the specified number.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dislikes",
                        "in": "query",
                        "description": "retrieve data by filtering the comments that have the amount of dislikes less than the specified number.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "retrieve data by filtering the comments that are made in the specified year.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Comments in JSON format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get a comment in a youtube video",
                "description": "retrieve a comment, given it existed, from a youtube video.",
                "operationId": "ca005bcffa22ab29b155dd7aa13b5d6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "video id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update a comment by id",
                "description": "update and store the updated data to a specific comment in the database. The comment must be the one made by your channel. Otherwise, an error will display.",
                "operationId": "CommentPut",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the comment from your channel. You could look through your comments made by your channel in the Channels tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "comment": {
                                        "description": "Update your comment body",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid video ID"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment in a video",
                "description": "store the data as a comment in a youtube video.",
                "operationId": "d448ba6444f0ff2770b763b441e91f6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "video id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "comment": {
                                        "description": "Post a new comment",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment Not Found"
                    },
                    "401": {
                        "description": "Unauthorised Access"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete a comment by its ID",
                "description": "Delete a comment specified by the ID parameter.",
                "operationId": "CommentDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/successful_delete"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/channel/comments": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get comment(s) made by your channel",
                "description": "retrieve comment(s), given it existed, from the token bearer.",
                "operationId": "ad365c6550bc8b4c8a4c048e7101697c",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "404": {
                        "description": "Comment Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "youtube_video": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                    },
                    "description": {
                        "type": "text",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque sed erat mi. Vivamus mollis orci vitae neque fringilla, non interdum urna commodo. Cras at blandit lorem, malesuada posuere felis."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-10-29"
                    },
                    "duration": {
                        "type": "integer",
                        "example": "30"
                    },
                    "likes": {
                        "type": "integer",
                        "example": "30"
                    },
                    "dislikes": {
                        "type": "integer",
                        "example": "30"
                    },
                    "views": {
                        "type": "integer",
                        "example": "30"
                    },
                    "link": {
                        "type": "text",
                        "example": "youtube.com\\/watch?v=1fa3d3b7-d0c7-3e5e-89e7-3713d0e74f87"
                    },
                    "thumbnail": {
                        "type": "text",
                        "example": "https:\\/\\/via.placeholder.com\\/640x480.png\\/005500?text=thumbnail+voluptatem"
                    },
                    "channel": {
                        "$ref": "#/components/schemas/Channel"
                    },
                    "comments": {
                        "$ref": "#/components/schemas/Comment"
                    }
                },
                "type": "object"
            },
            "Channel": {
                "properties": {
                    "id": {
                        "description": "channel id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "name of the channel",
                        "type": "string",
                        "example": "James"
                    },
                    "subscribers": {
                        "description": "number of subscribers",
                        "type": "integer",
                        "example": "1100"
                    },
                    "created_at": {
                        "description": "the date when the channel has been created",
                        "type": "string",
                        "format": "date",
                        "example": "2019-10-06"
                    },
                    "hasVideo": {
                        "description": "the amount of video",
                        "type": "integer",
                        "format": "integer",
                        "example": "100"
                    },
                    "videos": {
                        "$ref": "#/components/schemas/youtube_video"
                    }
                },
                "type": "object"
            },
            "successful_delete": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The data has been successfully deleted"
                    },
                    "status": {
                        "type": "integer",
                        "example": "202"
                    }
                },
                "type": "object"
            },
...
            "Comment": {
                "properties": {
                    "id": {
                        "description": "comment id",
                        "type": "integer",
                        "example": "1"
                    },
                    "text": {
                        "description": "comment texts",
                        "type": "text",
                        "example": "lorem ipsum"
                    },
                    "likes": {
                        "description": "number of likes",
                        "type": "integer",
                        "example": "1100"
                    },
                    "dislikes": {
                        "description": "number of dislikes",
                        "type": "integer",
                        "example": "90"
                    },
                    "commented_at": {
                        "description": "The date of the comment",
                        "type": "string",
                        "format": "string",
                        "example": "6 years ago"
                    },
                    "channel": {
                        "$ref": "#/components/schemas/Channel"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}